Behaviour based on gen_server, better results.
As of this commit it looks like the behaviour, so long as you stick to
it's public interface, is minimally workable at least. There's a CT
suite to help establish that but no documention or more elaborate
examples. The module itself still has some commented out functions,
etc, but it's coming along.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>